# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Project Backend API
  version: 1.0.0
  description: |
    Микросервисная система управления дефектами/заказами.
    Содержит 3 узла: API Gateway, Users Service, Orders Service.
    Все ответы имеют единый формат:
    - success: boolean
    - data: объект результата
    - error: { code, message }

servers:
  - url: http://localhost:4000/api/v1
    description: API Gateway (local)
  - url: http://api_gateway:4000/api/v1
    description: API Gateway (docker)

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: INVALID_INPUT
            message:
              type: string
              example: "Missing fields"

    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        roles:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string

    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        roles:
          type: array
          items:
            type: string

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            token:
              type: string
              example: "jwt-token"

    Profile:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            id:
              type: string
            email:
              type: string
            name:
              type: string
            roles:
              type: array
              items: { type: string }

    Order:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              qty:
                type: number
        total:
          type: number
        status:
          type: string
          enum: [created, in_progress, done, cancelled]
        createdAt:
          type: string
        updatedAt:
          type: string

    OrderCreateRequest:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              qty:
                type: number
        total:
          type: number

    OrderStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [created, in_progress, done, cancelled]

paths:

  /users/register:
    post:
      summary: Регистрация пользователя
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Успешная регистрация
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      id: { type: string }
                      email: { type: string }
        "409":
          description: Такой пользователь уже существует
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/login:
    post:
      summary: Вход пользователя
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Неверный логин/пароль
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/me:
    get:
      summary: Получение профиля
      security: [{ BearerAuth: [] }]
      tags: [Users]
      responses:
        "200":
          description: Профиль
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          description: Нет токена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users:
    get:
      summary: Получить список всех пользователей (admin)
      security: [{ BearerAuth: [] }]
      tags: [Users]
      responses:
        "200":
          description: Список пользователей
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
        "403":
          description: Нет прав
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders:
    get:
      summary: Список заказов текущего пользователя
      security: [{ BearerAuth: [] }]
      tags: [Orders]
      responses:
        "200":
          description: Список заказов
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Order"

    post:
      summary: Создать заказ
      security: [{ BearerAuth: [] }]
      tags: [Orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCreateRequest"
      responses:
        "200":
          description: Создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      summary: Получить конкретный заказ (если принадлежит пользователю)
      security: [{ BearerAuth: [] }]
      tags: [Orders]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Заказ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "403":
          description: Доступ запрещён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      summary: Отменить заказ
      security: [{ BearerAuth: [] }]
      tags: [Orders]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Отменён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}/status:
    patch:
      summary: Обновить статус заказа
      security: [{ BearerAuth: [] }]
      tags: [Orders]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Обновлено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"